переделала корзину на редакс

+++ src/components/Card/Card.jsx
@@ -13,38 +13,38 @@ import { openNotification } from '../Notification/Notification';
 
 export const Card = ({ product, pictures, name, wight, price, discount, onPlus }) => {
    const actualUser = useSelector((state) => state.user.data);
-   const [isAdded, setIsAdded] = useState(false);
+   // const [isAdded, setIsAdded] = useState(false);
    const dispatch = useDispatch();
    const navigate = useNavigate();
+   // console.log(product);
 
-   console.log(product);
+   const cartItem = useSelector((state) =>
+      state.basket.items.find((obj) => obj.id === product._id)
+   );
+   const addedCount = cartItem ? cartItem.count : 0;
+
+   // console.log(product);
 
    const isLiked = findLike(product, actualUser);
    const handleLikeClick = () => {
       dispatch(fetchChangeLikeProduct(product));
    };
 
-   // const areYouSure = () => {
-   //    if (!alert('Вы уверены что хотите удалить товар?')) {
-   //       return;
-   //    } else {
-   //       deleteCard(product._id);
-   //    }
-   // };
-
    const deleteCard = async (id) => {
       try {
-         alert('Вы уверены что хотите удалить товар?');
-         dispatch(fetchDeleteProducts(id));
-         navigate('/catalog');
-         openNotification('success', 'Успешно!', 'Ваш товар успешно удален');
+         if (window.confirm('Вы действительно хотите удалить товар?')) {
+            dispatch(fetchDeleteProducts(id));
+            navigate('/catalog');
+            openNotification('success', 'Успешно!', 'Ваш товар успешно удален');
+         }
       } catch (error) {
          openNotification('error', 'Ошибка!', 'Ваш товар не был удален');
       }
    };
+
    const onClickToBasket = () => {
       onPlus(product);
-      setIsAdded(!isAdded);
+      // setIsAdded(!isAdded);
    };
 
    return (
@@ -75,12 +75,15 @@ export const Card = ({ product, pictures, name, wight, price, discount, onPlus }
          <div className="card__buttons">
             <span
                onClick={() => onClickToBasket()}
-               className={`btn btn_type_primary ${
-                  isAdded ? 'btn_type_primary-active' : 'btn_type_primary'
-               }`}
+               className={`btn btn_type_primary`}
+               // className={`btn btn_type_primary ${
+               //    isAdded ? 'btn_type_primary-active' : 'btn_type_primary'
+               // }`}
             >
-               {isAdded ? <Link to={'/cart'}>В корзине</Link> : 'В корзину'}
+               {/* {isAdded ? <Link to={'/cart'}>В корзине</Link> : 'В корзину'} */}
+               {/* {isAdded ? 'В корзине' : 'В корзину'} */}В корзину
             </span>
+            {addedCount > 0 && <span className="added-count">{addedCount}</span>}
             {actualUser._id === product.author._id && (
                <Basket onClick={() => deleteCard(product._id)} className="card__basket" />
             )}
